Cleaned and optimized the whole CMake part
- Also, fixed some simple compilation warnings

+++ src/OmniVDescriptors.cpp
@@ -190,7 +190,7 @@ namespace OmniV {
 		for (auto& write : writes) {
 			write.dstSet = set;
 		}
-		vkUpdateDescriptorSets(pool.omnivDevice.device(), writes.size(), writes.data(), 0, nullptr);
+		vkUpdateDescriptorSets(pool.omnivDevice.device(), (uint32_t) writes.size(), writes.data(), 0, nullptr);
 	}
 
 }
